requirements.txt
Django==5.0.6
psycopg[binary]==3.1.19
django-phonenumber-field[phonenumbers]==7.3.0
